fixed bat I/O redirects, added copyrights, adjustments to library
