+++ bat/httplib/httplib.go
@@ -1,5 +1,6 @@
 // Copyright 2014 beego Author. All Rights Reserved.
 // Copyright 2015 bat authors
+// Modifications copyright 2018 ETH Zurich
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
