+++ lib/shttp/examples/minimal/client.go
@@ -1,3 +1,17 @@
+// Copyright 2018 ETH Zurich
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.package main
+
 package main
 
 import (
@@ -9,7 +23,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/netsec-ethz/scion-apps/lib/pathutil"
+	"github.com/netsec-ethz/scion-apps/lib/scionutil"
 	"github.com/netsec-ethz/scion-apps/lib/shttp"
 	"github.com/scionproto/scion/go/lib/snet"
 )
@@ -29,7 +43,9 @@ func main() {
 	}
 
 	if *interactive {
-		pathutil.ChoosePath(*interactive, "shortest", lAddr, rAddr)
+		scionutil.ChoosePathInteractive(lAddr, rAddr)
+	} else {
+		scionutil.ChoosePathByMetric(scionutil.Shortest, lAddr, rAddr)
 	}
 
 	// Make a map from URL to *snet.Addr
